+++ src/main/kotlin/io/openfuture/chain/domain/block/BlockCreationEvent.kt
@@ -1,0 +1,7 @@
+package io.openfuture.chain.domain.block
+
+import io.openfuture.chain.entity.transaction.BaseTransaction
+
+class BlockCreationEvent(
+    val pendingTransactions: MutableList<BaseTransaction>
+)