+++ src/main/kotlin/io/openfuture/chain/domain/block/Signature.kt
@@ -1,0 +1,6 @@
+package io.openfuture.chain.domain.block
+
+data class Signature(
+    val value: String,
+    val publicKey: String
+)