+++ src/main/kotlin/io/openfuture/chain/entity/transaction/BaseTransaction.kt
@@ -3,6 +3,7 @@ package io.openfuture.chain.entity.transaction
 import com.fasterxml.jackson.annotation.JsonIgnore
 import io.openfuture.chain.entity.base.BaseModel
 import io.openfuture.chain.entity.Block
+import io.openfuture.chain.entity.MainBlock
 import javax.persistence.*
 
 @Entity
@@ -37,6 +38,6 @@ abstract class BaseTransaction(
     @JsonIgnore
     @ManyToOne
     @JoinColumn(name = "block_id", nullable = true)
-    var block: Block? = null
+    var block: MainBlock? = null
 
 ) : BaseModel()