+++ src/main/resources/db/migration/V1_1__table_blocks.sql
@@ -1,11 +1,10 @@
 CREATE TABLE blocks (
-  id             INTEGER PRIMARY KEY,
-  height         BIGINT NOT NULL UNIQUE,
-  nonce          BIGINT NOT NULL,
-  timestamp      BIGINT NOT NULL UNIQUE,
-  merkle_hash    VARCHAR NOT NULL UNIQUE ,
-  hash           VARCHAR NOT NULL UNIQUE,
-  previous_hash  VARCHAR NOT NULL UNIQUE,
-  node_key       VARCHAR NOT NULL,
-  node_signature VARCHAR NOT NULL
+  id            INTEGER PRIMARY KEY,
+  hash          VARCHAR NOT NULL,
+  height        BIGINT NOT NULL,
+  previous_hash VARCHAR NOT NULL,
+  merkle_hash   VARCHAR NOT NULL,
+  timestamp     BIGINT  NOT NULL,
+  signature VARCHAR NOT NULL,
+  type_id       INTEGER NOT NULL
 );