+++ CHANGELOG.md
@@ -27,10 +27,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Connection establishing logic
 - Scheduled check of connections
 - Reconnect logic
+- Generating private and public key pairs on app start and saving in files if not exist
+- Shuffle active delegates
+- Transaction merkle hash calculation
+- Block production
+- Block validation
+- Converters for block, transaction, signature from prototype to business object and vice versa
+- Accumulation signatures from nodes
+- Block approving
+- Block applying by scheduling
 
 ### Changed
 - Default account info is added to the doGenerateMaster end-point response
 - doGenerateMaster end-point renamed to doRestore
+- Block separated by types: GenesisBlock and MainBlock
 
 ## [0.2.0] - 2018-06-29
 ### Added